run test starting server
