+++ package.json
@@ -10,7 +10,7 @@
     "start": "nodemon --exec babel-node src/index.js",
     "test:run-server": "TEST_DATABASE=mytestdatabase npm start",
     "test:execute-test": "mocha --require @babel/register 'src/**/*.spec.js'",
-    "test": "echo \"No test specified\" && exit 0"
+    "test": "set IS_TEST=true&& mocha --timeout 15000 --require @babel/register src/**/*.spec.js"
   },
   "keywords": [],
   "author": "Robin Wieruch <hello@rwieruch.com> (https://www.robinwieruch.de)",
@@ -38,7 +38,7 @@
     "graphql-iso-date": "^3.6.1",
     "graphql-resolvers": "^0.2.2",
     "jsonwebtoken": "^8.4.0",
-    "pg": "^7.6.1",
+    "tedious": "^3.0.1",
     "sequelize": "^4.41.2",
     "uuid": "^3.3.2"
   }
