+++ src/models/index.js
@@ -1,18 +1,23 @@
 import Sequelize from 'sequelize';
 
 let sequelize;
-if (process.env.DATABASE_URL) {
-  sequelize = new Sequelize(process.env.DATABASE_URL, {
+
+const { DATABASE_URL, IS_TEST, TEST_DATABASE, DATABASE, DATABASE_USER, DATABASE_PASSWORD, HOST } = process.env;
+
+if (DATABASE_URL) {
+  sequelize = new Sequelize(DATABASE_URL, {
     dialect: 'postgres',
   });
 } else {
+  const database = IS_TEST === 'true' ? TEST_DATABASE : DATABASE;
   sequelize = new Sequelize(
-    process.env.TEST_DATABASE || process.env.DATABASE,
-    process.env.DATABASE_USER,
-    process.env.DATABASE_PASSWORD,
+    database,
+    DATABASE_USER,
+    DATABASE_PASSWORD,
     {
-      dialect: 'postgres',
-    },
+      host: HOST,
+      dialect: 'mssql',  // postgres
+    }
   );
 }
